requirements.txt
bcrypt~=4.3.0
mysql-connector-python~=9.3.0
logging~=0.4.9.6
python-dotenv~=1.1.1